I-#128: Move context and hooks related to theme move to theme folder.
